==> code/prep/0-get_govinfo.Rmd <==
---
title: "Quick and dirty download of gov.info bulk bills data"
author: 
  - "Cole Tanigawa-Lau"
date: ""
output:
  html_document:
    df_print: paged
    toc: yes
    toc_depth: 3
    toc_float: yes
    theme: paper
    highlight: tango
editor_options:
  chunk_output_type: inline
params: 
  download: FALSE
  unzip: FALSE
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.width = 7, fig.height = 5)
proj_dir <- tryCatch(rprojroot::find_rstudio_root_file(),
                     error = function(e) rprojroot::find_root(".gitignore"))

knitr::opts_knit$set(root.dir = proj_dir)
```

```{r about, echo=FALSE, results = "asis"}
cat(
  sprintf("Updated: %s.", 
          format(Sys.time(), "%b %d, %Y at %H:%M:%S", usetz = TRUE)),
  sprintf("Working directory: `%s`.", getwd()),
  sep = "\n\n"
)
```


Reads:

  - FILEPATH

Writes:

  - Unzips to /scratch/billtag/data/legislation/govinfo

# Setup

```{r packages, warning=FALSE, message=FALSE}
rm(list = setdiff(ls(), "params"))

library(arrow)
library(qs)
library(data.table)
library(dplyr)
library(stringr)
library(purrr)
library(xml2)

source("code/functions.R")

scratch_dir <- "/scratch/users/coletl/billtag/data/legislation/govinfo"

# Overwrite option when unzipping
overwrite <- FALSE
```

# Construct download URLS

```{r}
bulk_stems <- 
  expand.grid(congress = 113:117,
              session = 1:2,
              bill_type = c("hconres", "hjres", "hr", "hres", 
                            "s", "sconres", "sjres", "sres")) %>% 
  as.data.table()

bulk_stems[ , root := "https://www.govinfo.gov/bulkdata/BILLS"]

bulk_stems[ , zip_url := 
             str_glue("{root}/{congress}/{session}/{bill_type}/BILLS-{congress}-{session}-{bill_type}.zip",
                      .envir = bulk_stems)
           ][ ,
              file_name := str_glue("{congress}-{session}-{bill_type}.zip", 
                                    .envir = bulk_stems)]
```

Construct zip file paths to write to.
```{r}
bulk_stems[ , dir := str_glue("data/legislation/govinfo/{congress}",
                              .envir = bulk_stems)
            ][ , path := file.path(dir, file_name)]

```

# Download
```{r download, eval = params$download}
bulk_stems[ , map(unique(dir), dir.create)] %>% 
  invisible()

bulk_stems[ , 
            map2(zip_url, path, 
                 ~ download.file(url = .x, destfile = .y))
            ]
```

Unzip.
```{r unzip, eval = params$unzip}
system.time(
  map(bulk_stems$path, 
      ~ unzip(.x, exdir = scratch_dir, overwrite = overwrite) %>% 
        # suppress warnings about not overwriting file
        suppressWarnings()
      )
  )

length(list.files(scratch_dir))
```

# Convert to nested list
```{r, eval = params$download}
xml_files <-
  list.files(path = scratch_dir, pattern = ".xml",
             recursive = TRUE, full = TRUE)

lxml <- list()

system.time(
  for (file in xml_files) {
    lxml[file] <- 
      tryCatch(read_xml(file) %>% as_list(),
               error = function(e) sprintf("ERROR: %s", e$message))
  }
)


qsave(lxml, "data/legislation/govinfo/bills_xml.qs")
```


```{r read_xml_list, eval = !params$download}
lxml <- qread("data/legislation/govinfo/bills_xml.qs")
```

Good error rate.
```{r}
err_lgl <- map_lgl(lxml, ~ class(.x) == "character")
sum(err_lgl); mean(err_lgl)

stopifnot(sum(err_lgl) < 100)
stopifnot(mean(err_lgl) < 1e-4)

lxml <- lxml[ - which(err_lgl) ]
```

# Get text from XML list

```{r get_text_xml}
bill_tags <- 
    str_extract(names(lxml), "(?<=BILLS-)[0-9a-z]+(?=\\.xml)") %>%
    gsub("[a-z]+\\d*$", "", .)


# Convert format from 113hconres100 to hr2596-115
bill_ids <-
    gsub("(\\d+)([a-z]+)(\\d+)", "\\2\\3-\\1", bill_tags)

bill_text <- 
    data.table(id = bill_ids,
               text = vapply(lxml, get_text_govinfo, 
                                  FUN.VALUE = character(1)))

bill_text[ , text := gsub("\n\t\t\t", "", tolower(text), fixed = TRUE)]
bill_text <- data.table:::unique.data.table(bill_text)
```


Still some duplicate bill IDs remaining. Keep whichever has the longer text.
```{r}
bill_text[ , nchar := nchar(text)]

bills_final <- bill_text[order(nchar), .(text = last(text)), by = id]

stopifnot(anyDuplicated(bills_final$id) == 0)

round(nrow(bills_final) / nrow(bill_text), 3) * 100
```


# Export
```{r}
write_parquet(bills_final, "data/legislation/govinfo/bill_text.parquet")

set.seed(575)
bill_samp <- dplyr::sample_n(bills_final, 100)
write_parquet(bill_samp, "data/legislation/govinfo/bill_text_samp.parquet")
```
